+++ vitest.config.ts
@@ -14,8 +14,10 @@ export default defineConfig({
 		environment: "node",
 		coverage: {
 			enabled: true,
+			all: true,
 			provider: "istanbul",
-			reportsDirectory: "coverage"
+			reportsDirectory: "coverage",
+			exclude: ["scripts/**/*"]
 		}
 	}
 });